feat: 更新课程相关DTO和服务，添加参数校验和异常处理逻辑

+++ xuecheng-plus-content/xuecheng-plus-content-service/src/main/java/com/xuecheng/content/service/impl/CourseBaseServiceImpl.java
@@ -166,7 +166,7 @@ public class CourseBaseServiceImpl implements CourseBaseInfoService {
         }
 
         if (charge.equals("201001")) {
-            if (courseMarket.getPrice() == null || courseMarket.getPrice() <= 0) {
+            if (courseMarket.getPrice() == null || courseMarket.getPrice().floatValue() <= 0) {
                 throw new XueChengPlusException("课程价格不能为空");
             }
         }
